debug45

+++ lib/snips_json.c
@@ -56,7 +56,7 @@ void json_snips_log(char *parent, int count, ...) {
         if (curCount < 1)
         fprintf(jsonf, "\n\t\t\t\t\"%s\": %d%c", name, *((int *)value), count > 1 ? ',' : ' ');
         else
-        fprintf(jsonf, "\n\t\t\t\t\"%s\": %.9f%c", name, *((float *)value), count > 1 ? ',' : ' ');
+        fprintf(jsonf, "\n\t\t\t\t\"%s\": %.9f%c", name, *((int *)value), count > 1 ? ',' : ' ');
 
     } else {
       fprintf(jsonf, "\n\t\t\t\t\"%s\": null%c", name, count > 1 ? ',' : ' ');
